realtylens-airflow-dbt/requirements.txt
dbt-snowflake==1.6.0
apache-airflow-providers-snowflake==5.1.1
astronomer-cosmos==1.9.0
